Add support for NationalGrid NY Metro
