+++ .gitignore
@@ -128,3 +128,7 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# Jetbrains IDEs
+.idea/
+*.iml
